+++ src/HealthChecks.Rabbitmq/HealthChecks.Rabbitmq.csproj
@@ -5,6 +5,7 @@
     <PackageTags>HealthCheck;Health;RabbitMQ</PackageTags>
     <Description>HealthChecks.RabbitMQ is the health check package for RabbitMQ.</Description>
     <Version>$(HealthCheckRabbitMQ)</Version>
+     <RootNamespace>HealthChecks.RabbitMQ</RootNamespace> <!--For backward naming compatibility-->
   </PropertyGroup>
 
   <ItemGroup>
