+++ test/UnitTests/DependencyInjection/IdSvr/IdSvrUnitTests.cs
@@ -1,0 +1,47 @@
+﻿using FluentAssertions;
+using HealthChecks.IdSvr;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Diagnostics.HealthChecks;
+using Microsoft.Extensions.Options;
+using System;
+using System.Linq;
+using Xunit;
+
+namespace UnitTests.HealthChecks.DependencyInjection.IdSvr
+{
+    public class idsvr_registration_should
+    {
+        [Fact]
+        public void add_health_check_when_properly_configured()
+        {
+            var services = new ServiceCollection();
+            services.AddHealthChecks()
+                .AddIdentityServer(new Uri("http://myidsvr"));
+
+            var serviceProvider = services.BuildServiceProvider();
+            var options = serviceProvider.GetService<IOptions<HealthCheckServiceOptions>>();
+
+            var registration = options.Value.Registrations.First();
+            var check = registration.Factory(serviceProvider);
+
+            registration.Name.Should().Be("idsvr");
+            check.GetType().Should().Be(typeof(IdSvrHealthCheck));
+        }
+        [Fact]
+        public void add_named_health_check_when_properly_configured()
+        {
+            var services = new ServiceCollection();
+            services.AddHealthChecks()
+                .AddIdentityServer(new Uri("http://myidsvr"), name: "my-idsvr-group");
+
+            var serviceProvider = services.BuildServiceProvider();
+            var options = serviceProvider.GetService<IOptions<HealthCheckServiceOptions>>();
+
+            var registration = options.Value.Registrations.First();
+            var check = registration.Factory(serviceProvider);
+
+            registration.Name.Should().Be("my-idsvr-group");
+            check.GetType().Should().Be(typeof(IdSvrHealthCheck));
+        }
+    }
+}
